+++ NEWS.md
@@ -1,6 +1,7 @@
 # landscapemetrics 0.3.2
 * Bugfixes
     * Bugfix in `check_landscape()` if NA values were present
+    * Bugfix in `pad_raster()` that pad_raster_cells > 1 did not work
 * Improvements
     * Removing all functions from `dplyr`
     * Better structure for helping functions of `check_landscape()`